Lazily validate gemspecs on error, instead of in advance
Most of the time gemspecs will be correct - particularily those from a remote source
The odd time, a gemspec is invalid and weird issues start arising. To combat this
a validation was added on load. For most applications, this is an insignificant side effect,
but for larger applications this can add upwards of 200-300ms on every boot.

Instead of validating on load, validate on error. That is, we can catch an error in the resolver
and validate the gemspecs there instead. If one is invalid, then raise that error. If none are invalid,
raise the original error. We assume here that an invalid gemspec is a cause for the error. Regardless of
whether or not this is true, we can assume that the invalid gemspec should be ruled out and an error
would be raised on the next run.